bitcoin/ai_crypto_trading_system/requirements.txt
# 核心依赖
pandas>=2.0.0
numpy>=1.24.0
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0
websockets>=11.0.0

# 交易所API
ccxt>=4.0.0
python-binance>=1.0.0

# AI模型接口
openai>=1.0.0
requests>=2.31.0

# 数据处理和机器学习
scikit-learn>=1.3.0
ta>=0.10.0  # 技术分析指标

# 配置和日志
pyyaml>=6.0
colorlog>=6.7.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.12.0

# Web界面
flask>=2.3.0
flask-socketio>=5.3.0

# 监控和指标
prometheus-client>=0.17.0
psutil>=5.9.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 开发工具
black>=23.0.0
flake8>=6.0.0